requirements.txt
flask
pytesseract
pillow
python-Levenshtein
scikit-learn
PyPDF2
python-docx
matplotlib
seaborn
pandas
pdfminer
fuzzywuzzy
pdfminer.six
PyMuPDF
pdfplumber
numpy
scipy
nltk
werkzeug
docx2txt
opencv-python
python-docx
pdfkit
reportlab
fpdf
openpyxl
pdf2image
sentence_transformers
pytesseract
easyocr